+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * fix(`ObjectToObjectMetadataFactory`): Remove remnants.
 * fix(`SubMapperInterface`): Context is now null by default.
+* dx(`SubMapperInterface`): `map()` now accepts null input, that will return
+  null.
 
 ## 0.6.4